Add sqlite db sink, migrate from Double to Decimal due to double precision causing flakely tests on trivial examples

+++ sqrl-engine-flink-connectors/sqrl-engine-flink-sqlite/src/main/java/com/datasqrl/flink/connectors/jdbc/dialect/sqlite/SqliteDDLFactory.java
@@ -73,7 +73,7 @@ public class SqliteDDLFactory implements JdbcDDLFactory {
       case DECIMAL:
       case FLOAT:
       case DOUBLE:
-        return "DOUBLE";
+        return "NUMERIC";
       case DATE:
         return "DATE";
       case TIME:
